Move top-level custom types module to adapters
