+++ dj_link/adapters/datajoint/__init__.py
@@ -1,1 +1,4 @@
 """Contains code initializing the adapters."""
+from .custom_types import PrimaryKey
+
+__all__ = ["PrimaryKey"]
