+++ dj_link/adapters/datajoint/gateway.py
@@ -5,13 +5,13 @@ from collections import defaultdict
 from itertools import groupby
 from typing import Iterable
 
-from dj_link.custom_types import PrimaryKey
 from dj_link.entities.custom_types import Identifier
 from dj_link.entities.link import Link, create_link
 from dj_link.entities.state import Commands, Processes, Update
 from dj_link.use_cases.gateway import LinkGateway
 
 from ...entities.state import Components
+from .custom_types import PrimaryKey
 from .facade import DJAssignments, DJLinkFacade, DJProcess
 from .identification import IdentificationTranslator
